+++ components/Searchbar/index.tsx
@@ -1,119 +1,21 @@
-import { useState, ReactNode, Dispatch, SetStateAction, useMemo } from "react";
+import { useState, Dispatch, SetStateAction, useMemo } from "react";
 import InputAdornment from "@mui/material/InputAdornment";
 import TextField, { TextFieldProps } from "@mui/material/TextField";
 import Search from "@mui/icons-material/Search";
 import Tune from "@mui/icons-material/Tune";
-import ExpandMore from "@mui/icons-material/ExpandMore";
 import IconButton from '@mui/material/IconButton';
 import Popover from '@mui/material/Popover';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import Box from "@mui/system/Box";
-import Accordion from "@mui/material/Accordion";
-import Checkbox from "@mui/material/Checkbox";
-import AccordionSummary from "@mui/material/AccordionSummary";
-import AccordionDetails from "@mui/material/AccordionDetails";
-import FormControlLabel from "@mui/material/FormControlLabel";
 import { Region } from "constants/Donation";
 
-import { DateRange as DateRangeComponent, Range } from 'react-date-range';
-import 'react-date-range/dist/styles.css'; // main style file
-import 'react-date-range/dist/theme/default.css'; // theme css file
-import Autocomplete, { AutocompleteProps } from "@mui/material/Autocomplete";
-import Chip from "@mui/material/Chip";
 import Badge from "@mui/material/Badge";
-import Cancel from "@mui/icons-material/Cancel";
 import Settings from "@mui/icons-material/Settings";
-
-type DynamicMultiSelectFilterProps<T> = {
-    values: T[],
-    setValues: Dispatch<SetStateAction<T[]>>
-    valueSet: T[],
-    renderValue?: (v: T) => ReactNode
-} & Omit<AutocompleteProps<T, true, false, false>, 'options' | 'renderInput'>;
-
-function DynamicMultiSelectFilter<T>({ values, setValues, valueSet, renderValue, ...props }: DynamicMultiSelectFilterProps<T>) {
-    const _renderValue = renderValue || ((v: T) => v + "");
-    const chips = values.map((v, i) => {
-        return (
-            <Chip key={i} label={_renderValue(v)}
-                deleteIcon={<Cancel/>} sx={{ marginRight: 1, marginTop: 1 }}
-                onDelete={() => setValues(val => val.filter(ov => ov !== v))}/>
-        );
-    });
-    return (
-        <Box sx={{ marginBottom: 2, display: 'flex', flexDirection: 'column' }}>
-            <Autocomplete<T, true, false, false>
-                {...props}
-                multiple
-                renderInput={(params) => <TextField {...params}  />}
-                options={valueSet}
-                value={values}
-                renderTags={() => ''} // disable rendering of tags into the textfield
-                onChange={(_, nv) => setValues(nv)} />
-            {/* Cool css trick to force overflow */}
-            <Box sx={{ width: 0, minWidth: '100%' }}>
-                {chips}
-            </Box>
-        </Box>
-    );
-}
-
-interface MultiSelectFilterProps<T> {
-    values: T[],
-    setValues: Dispatch<SetStateAction<T[]>>
-    valueSet: T[],
-    renderValue?: (v: T) => ReactNode
-}
-
-function MultiSelectFilter<T>({ values, setValues, valueSet, renderValue }: MultiSelectFilterProps<T>) {
-    const addOrRmv = (val: T, checked: boolean) => {
-        if (checked) {
-            setValues(vs => [...vs, val]);
-        } else {
-            setValues(vs => vs.filter(v => v !== val));
-        }
-    };
-    const _renderValue = renderValue || ((v: T) => v + "");
-
-    return (
-        <Box sx={{ display: 'flex', flexDirection: 'column' }}>
-            {valueSet.map((v, i) => {
-                return <FormControlLabel
-                    control={<Checkbox checked={values.find((ov) => v === ov) !== undefined}
-                    onChange={(_, c) => addOrRmv(v, c)} />}
-                    label={_renderValue(v)} key={i}/>
-            })}
-        </Box>
-    );
-}
-
-interface FilterContainerProps {
-    title: string,
-    children: ReactNode
-}
-
-const FilterContainer = ({title, children}: FilterContainerProps) => {
-    return (
-        <Accordion disableGutters elevation={0} sx={{
-            borderTop: `1px solid rgba(0, 0, 0, 0.12)`,
-            '&:before': { backgroundColor: 'unset' } // remove another border
-        }}>
-            <AccordionSummary
-                expandIcon={<ExpandMore/>}>
-                {title}
-            </AccordionSummary>
-            <AccordionDetails sx={{ paddingLeft: 2, paddingRight: 2, paddingTop: 0, paddingBottom: 0 }}>
-                {children}
-            </AccordionDetails>
-        </Accordion>
-    );
-}
-
-export interface DateRange {
-    start?: Date,
-    end?: Date
-}
+import { DynamicMultiSelectFilter } from "./DynamicMultiSelectFilter";
+import { MultiSelectFilter } from "./MultiSelectFilter";
+import { DateRangeFilter, DateRange } from "./DateRangeFilter";
+import { FilterContainer } from "./FilterContainer";
 
 type SearchbarProps = {
     tags: string[],
@@ -143,19 +45,6 @@ const Searchbar = ({
     }: SearchbarProps) => {
 
     const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
-
-    const range =  useMemo<Range[]>(() => {
-        // ref: https://github.com/hypeserver/react-date-range/issues/330#issuecomment-802601417
-        return selectedDateRange === undefined ? [{
-            startDate: undefined,
-            endDate: new Date(""),
-            key: 'selection'
-        }] : [{
-            startDate: selectedDateRange.start,
-            endDate: selectedDateRange.end,
-            key: 'selection'
-        }];
-    }, [selectedDateRange]);
 
     const state = useMemo(() => {
         return {
@@ -238,7 +127,7 @@ const Searchbar = ({
                         <MultiSelectFilter valueSet={Object.values(Region)} values={selectedRegions} setValues={setSelectedRegions} />
                     </FilterContainer>
                     <FilterContainer title="Date">
-                        <DateRangeComponent editableDateInputs={true} onChange={nr => setSelectedDateRange({ start: nr.selection?.startDate, end: nr.selection?.endDate })} ranges={range} moveRangeOnFirstSelection={true} />
+                        <DateRangeFilter value={selectedDateRange} setValue={setSelectedDateRange} />
                     </FilterContainer>
                     <FilterContainer title="Tags">
                         <DynamicMultiSelectFilter valueSet={tags} values={selectedTags} setValues={setSelectedTags} />
